requirements.txt
anyio==3.7.1
asgiref==3.8.1
bcrypt==3.2.0
cffi==1.16.0
click==8.1.7
dnspython==2.1.0
fastapi==0.70.0
gunicorn==20.1.0
h11==0.14.0
idna==3.7
motor==3.5.1
passlib==1.7.4
pycparser==2.22
pydantic==1.10.17
PyJWT==2.3.0
pymongo==4.8.0
python-decouple==3.3
six==1.16.0
sniffio==1.3.1
starlette==0.16.0
typing_extensions==4.12.2
uvicorn==0.14.0
